safe refactoring (#2552)
